Move `<CardMedia />` to `card-media/` subfolder
