+++ packages/components/src/card/index.js
@@ -3,3 +3,4 @@ export { default as CardBody } from './card-body';
 export { default as CardDivider } from './card-divider';
 export { default as CardFooter } from './card-footer';
 export { default as CardHeader } from './card-header';
+export { default as CardMedia } from './card-media';
