+++ packages/components/src/card/stories/media.js
@@ -7,8 +7,7 @@ import styled from '@emotion/styled';
 /**
  * Internal dependencies
  */
-import { Card, CardBody, CardFooter, CardHeader } from '../';
-import CardMedia from '../media';
+import { Card, CardBody, CardFooter, CardHeader, CardMedia } from '../';
 import { getCardStoryProps } from './_utils';
 
 export default { title: 'Components/Card/Media', component: CardMedia };
