+++ packages/components/src/card/test/index.js
@@ -6,8 +6,7 @@ import { mount, shallow } from 'enzyme';
 /**
  * Internal dependencies
  */
-import { Card, CardBody, CardDivider, CardFooter, CardHeader } from '../';
-import CardMedia from '../media';
+import { Card, CardBody, CardDivider, CardFooter, CardHeader, CardMedia } from '../';
 
 describe( 'Card', () => {
 	describe( 'basic rendering', () => {
