+++ packages/components/src/index.js
@@ -25,8 +25,7 @@ export { default as BaseControl } from './base-control';
 export { default as __experimentalBoxControl } from './box-control';
 export { default as Button } from './button';
 export { default as ButtonGroup } from './button-group';
-export { Card, CardBody, CardDivider, CardFooter, CardHeader } from './card';
-export { default as CardMedia } from './card/media';
+export { Card, CardBody, CardDivider, CardFooter, CardHeader, CardMedia } from './card';
 export { default as CheckboxControl } from './checkbox-control';
 export { default as ClipboardButton } from './clipboard-button';
 export { default as __experimentalColorEdit } from './color-edit';
